moved case details into own components folder
