+++ app/case-studies/[slug]/page.tsx
@@ -1,11 +1,11 @@
-import CaseDetailsDescription from "@/components/CaseDetailsDescription";
-import CaseDetailsFigma from "@/components/CaseDetailsFigma";
-import CaseDetailsHeader from "@/components/CaseDetailsHeader";
-import CaseDetailsLearnings from "@/components/CaseDetailsLearnings";
-import CaseDetailsOtherStudies from "@/components/CaseDetailsOtherStudies";
-import CaseDetailsProblemStatement from "@/components/CaseDetailsProblemStatement";
-import CaseDetailsProcess from "@/components/CaseDetailsProcess";
-import CaseDetailsTechStack from "@/components/CaseDetailsTechStack";
+import CaseDetailsDescription from "@/components/case-details/CaseDetailsDescription";
+import CaseDetailsFigma from "@/components/case-details/CaseDetailsFigma";
+import CaseDetailsHeader from "@/components/case-details/CaseDetailsHeader";
+import CaseDetailsLearnings from "@/components/case-details/CaseDetailsLearnings";
+import CaseDetailsOtherStudies from "@/components/case-details/CaseDetailsOtherStudies";
+import CaseDetailsProblemStatement from "@/components/case-details/CaseDetailsProblemStatement";
+import CaseDetailsProcess from "@/components/case-details/CaseDetailsProcess";
+import CaseDetailsTechStack from "@/components/case-details/CaseDetailsTechStack";
 import ContactBanner from "@/components/ContactBanner";
 import { client, urlFor } from "@/lib/sanity";
 import { Project } from "@/types";
