+++ components/case-details/CaseDetailsFigma.tsx
